requirements.txt
numpy==1.16.2
torch==1.2.0
torchvision==0.5.0
pyyaml
opencv-contrib-python
